coreset prototype

+++ coreset/utils.py
@@ -107,11 +107,7 @@ def craig_baseline(data, K):
     # features = data.astype(np.float32)
     features = data.astype(np.single)
     # D = pairwise_distances(features, metric="euclidean", n_jobs=20).astype(np.single)
-    D = pairwise_distances(
-        features.astype(np.int32),
-        metric="euclidean",
-        n_jobs=-1,
-    ).astype(np.half)
+    D = pairwise_distances(features, metric="euclidean", n_jobs=-1).astype(np.half)
     D = D.max() - D
     V = np.arange(len(features), dtype=int).reshape(-1, 1)
     locator = FacilityLocation(D=D, V=V)
